Create README.md
I make a project using Arduino, Ultrasonic sensor and 1Sheeld that allows you to tap into your smartphone's sensors and capabilities and lets you use them in your Arduino projects. Basically, It is a hardware shield that sits on top of your Arduino board and communicates over Bluetooth to 1Sheeld app, to transfer data between it and your smartphone.